Centralize types, fix circular dependencies
