+++ src/index.ts
@@ -1,3 +1,3 @@
 import FragmentPlugin from './SwupFragmentPlugin.js';
 export default FragmentPlugin;
-export type { Options, Rule, FragmentElement } from './SwupFragmentPlugin.js';
+export type { Options, Rule, FragmentElement } from './inc/types.js';
